requirements.txt
jax[tpu]
flax
tensorflow-cpu
tensorflow_datasets
